Research branch for mbedTLS
We have hitherto used GnuTLS as our driver for STARTTLS logic.
It has been reasonably flexible, though its uptake of our work
on TLS-KDH has been postponed a lot, mostly due to the work on
TLS version 1.3.  We are interested in offering choice anyway,
and also need to push TLS-KDH forward (it's been years since
we developed it) and see a good option in mbedTLS.

Unlike GnuTLS, which is known for its reasonably extensive
set of features, mbedTLS tends to be trimmed-down.  It will
be really interesting to see if it can be, ehm, embedded into
our TLS Pool.  Even for users who might not use it directly,
it could be an interesting option in catastrophic situations
to be able to switch back to this highly conservative driver.

We may not be able to replace all the functionality that
GnuTLS offered with mbedTLS, but the differences in user
experience ought to be negligable.