Research branch for a STARTSSH method
Most protocols use a STARTTLS method to initiate TLS over a
formerly plaintext connection.  There is no reason why there
could not be similar mechanisms based on other secure
protocol, presumably with somewhat different characteristics.
This research branch considers a STARTSSH method based on SSH.

The OpenSSH protocol is a secure layer with a very specific
purpose, namely shell interactions, which has been expanded with
facilities such as connection forwarding and file exchange.  But
there is no reason why the carrier could not be used to do the
things we do with TLS today.  And it is good to have alternatives.

This project is about exploring those options.  And how they might
integrate with OpenSSH.  And with end-user tools, such as browsers
and remote file systems, remote desktops, ...